tool_load and tool_use scripts added
tool_load is just a script used to make new tools that are loaded when the game is launched
tool_use is a script that allows you to pass two arguments into it (tool id, direction of use) and, depending on the tool id passed, will go through a switch/case statement that then executes some action based on the tool id

as of right now, there are four tools created with tool_load. these tool's actions can be found in tool_use. right now, all of the tools simply create a particle (this particle is created at a different distance away from the player depending on which tool you're using and at a different direction depending on the direction passed into tool_use)